refactor: format indents, rename module
