+++ cognee/infrastructure/databases/exceptions/EmbeddingException.py
